extract tests to a module
